+++ lib/v2/datasource/local/qr_code_service.dart
@@ -1,7 +1,5 @@
 import 'package:async/async.dart';
 // ignore: import_of_legacy_library_into_null_safe
-import 'package:dart_esr/dart_esr.dart';
-import 'package:seeds/v2/datasource/local/models/scan_qr_code_result_data.dart';
 import 'package:seeds/v2/datasource/local/util/seeds_esr.dart';
 
 class QrCodeService {
@@ -14,26 +12,9 @@ class QrCodeService {
     }
 
     final SeedsESR esr = SeedsESR(uri: scanResult);
-    return esr.resolve(account: accountName).then((value) => processResolvedRequest(esr)).catchError((onError) {
+    return esr.resolve(account: accountName).then((value) => esr.processResolvedRequest()).catchError((onError) {
       print(" processQrCode : Error processing QR code");
       return ErrorResult("Error processing QR code");
     });
   }
-}
-
-Result processResolvedRequest(SeedsESR esr) {
-  final Action action = esr.actions.first;
-  if (_canProcess(action)) {
-    final Map<String, dynamic> data = Map<String, dynamic>.from(action.data! as Map<dynamic, dynamic>);
-    print(
-        " processResolvedRequest: Success QR contract: ${action.account} action: ${action.name} data: ${action.data!}");
-    return ValueResult(ScanESRResultData(data: data, accountName: action.account, actionName: action.name));
-  } else {
-    print("processResolvedRequest: canProcess is false: ");
-    return ErrorResult("Invalid QR code");
-  }
-}
-
-bool _canProcess(Action action) {
-  return action.account!.isNotEmpty && action.name!.isNotEmpty;
 }
